mi_proyecto/requirements.txt
Django==3.2.2
mysql-connector-python
mysqlclient